+++ Libplanet.Action.Tests/ActionContextTest.cs
@@ -1,4 +1,3 @@
-using Libplanet.Action.State;
 using Libplanet.Action.Tests.Mocks;
 using Libplanet.Crypto;
 using Libplanet.Types.Blocks;
@@ -36,7 +35,7 @@ namespace Libplanet.Action.Tests
                     miner: _address,
                     blockIndex: 1,
                     blockProtocolVersion: Block.CurrentProtocolVersion,
-                    previousState: Account.Create(MockAccountState.Empty),
+                    previousState: MockAccount.Empty,
                     randomSeed: seed,
                     gasLimit: 0
                 );
@@ -54,7 +53,7 @@ namespace Libplanet.Action.Tests
                 miner: _address,
                 blockIndex: 1,
                 blockProtocolVersion: Block.CurrentProtocolVersion,
-                previousState: Account.Create(MockAccountState.Empty),
+                previousState: MockAccount.Empty,
                 randomSeed: 0,
                 gasLimit: 0
             );
@@ -65,7 +64,7 @@ namespace Libplanet.Action.Tests
                 miner: _address,
                 blockIndex: 1,
                 blockProtocolVersion: Block.CurrentProtocolVersion,
-                previousState: Account.Create(MockAccountState.Empty),
+                previousState: MockAccount.Empty,
                 randomSeed: 0,
                 gasLimit: 0
             );
@@ -76,7 +75,7 @@ namespace Libplanet.Action.Tests
                 miner: _address,
                 blockIndex: 1,
                 blockProtocolVersion: Block.CurrentProtocolVersion,
-                previousState: Account.Create(MockAccountState.Empty),
+                previousState: MockAccount.Empty,
                 randomSeed: 1,
                 gasLimit: 0
             );
@@ -112,7 +111,7 @@ namespace Libplanet.Action.Tests
                     miner: _address,
                     blockIndex: 1,
                     blockProtocolVersion: Block.CurrentProtocolVersion,
-                    previousState: Account.Create(MockAccountState.Empty),
+                    previousState: MockAccount.Empty,
                     randomSeed: i,
                     gasLimit: 0
                 );
@@ -132,7 +131,7 @@ namespace Libplanet.Action.Tests
                 miner: _address,
                 blockIndex: 1,
                 blockProtocolVersion: Block.CurrentProtocolVersion,
-                previousState: Account.Create(MockAccountState.Empty),
+                previousState: MockAccount.Empty,
                 randomSeed: _random.Next(),
                 gasLimit: 0);