+++ Libplanet.Tests/Store/StateStoreTracker.cs
@@ -22,10 +22,24 @@ namespace Libplanet.Tests.Store
             return _stateStore.GetStateRoot(stateRootHash, readOnly);
         }
 
+        public IRecordableTrie GetRecordableStateRoot(
+            HashDigest<SHA256>? stateRootHash,
+            bool readOnly = false)
+        {
+            Log(nameof(GetRecordableStateRoot), stateRootHash);
+            return _stateStore.GetRecordableStateRoot(stateRootHash, readOnly);
+        }
+
         public void PruneStates(IImmutableSet<HashDigest<SHA256>> survivingStateRootHashes)
         {
             Log(nameof(PruneStates), survivingStateRootHashes);
             _stateStore.PruneStates(survivingStateRootHashes);
+        }
+
+        public IRecordableTrie CastToRecordableTrie(ITrie trie)
+        {
+            Log(nameof(CastToRecordableTrie), trie.Hash);
+            return _stateStore.CastToRecordableTrie(trie);
         }
 
         public void Dispose()